pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "chronoface"
version = "0.1.0"
description = "Chronoface local face clustering and collage generator"
authors = [{name = "Chronoface contributors"}]
license = {text = "MIT"}
requires-python = ">=3.11"
readme = "README.md"
keywords = ["photos", "faces", "clustering", "collage"]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.11",
  "License :: OSI Approved :: MIT License",
  "Framework :: FastAPI"
]
dependencies = [
  "fastapi~=0.110",
  "uvicorn[standard]~=0.29",
  "opencv-contrib-python~=4.9",
  "onnxruntime~=1.17",
  "scikit-learn~=1.4",
  "pillow>=10.1",
  "pillow-heif~=0.16",
  "piexif~=1.1",
  "imagehash~=4.3",
  "pydantic~=2.6",
  "python-multipart~=0.0.9",
  "loguru~=0.7",
  "typer~=0.12",
  "numpy>=1.26,<3",
  "orjson~=3.10"
]

[project.optional-dependencies]
dev = [
  "ruff~=0.4",
  "black~=24.4",
  "mypy~=1.8",
  "pytest~=8.1",
  "pytest-asyncio~=0.23",
  "httpx~=0.27",
  "respx~=0.20"
]

[project.urls]
Homepage = "https://github.com/chronoface/chronoface"
Repository = "https://github.com/chronoface/chronoface"

[tool.setuptools]
packages = ["backend"]

[tool.black]
line-length = 88
target-version = ["py311"]

[tool.ruff]
line-length = 88
target-version = "py311"
select = ["E", "F", "I", "B", "UP", "RUF"]
ignore = ["E203", "E266"]

[tool.mypy]
python_version = "3.11"
packages = ["backend"]
strict = true
warn_unused_configs = true
warn_return_any = true
warn_unused_ignores = true
show_error_codes = true
plugins = []

[tool.pytest.ini_options]
minversion = "8.0"
addopts = "-ra -q"
testpaths = ["backend/tests"]
